requirements.txt
altgraph==0.17.4
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
flatbuffers==24.3.25
humanfriendly==10.0
idna==3.8
imageio==2.35.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
llvmlite==0.43.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
onnxruntime==1.19.0
opencv-python-headless==4.10.0.84
packaging==24.1
pefile==2024.8.26
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
protobuf==5.28.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyMatting==1.1.12
pyreadline3==3.4.1
pywin32-ctypes==0.2.3
referencing==0.35.1
rembg==2.0.59
requests==2.32.3
rpds-py==0.20.0
scikit-image==0.24.0
scipy==1.14.1
setuptools==74.0.0
sympy==1.13.2
tifffile==2024.8.30
tk==0.1.0
tqdm==4.66.5
ttkthemes==3.2.2
urllib3==2.2.2
